src/serious_python/example/searxng_example/app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.11.12
click==8.3.0
colorama==0.4.6
fasttext-predict==0.9.2.4
Flask[async]==3.1.2
flask-babel==4.0.0
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx[http2]==0.28.1
httpx-socks[asyncio]==0.10.1
hyperframe==6.1.0
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp-utils==1.0.0
mdurl==0.1.2
msgspec==0.20.0
multidict==6.7.0
platformdirs==4.5.0
propcache==0.4.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-socks==2.7.2
pytz==2025.2
PyYAML==6.0.3
queuelib==1.8.0
requests==2.32.5
rich==14.2.0
searxng @ file:pypackages/searxng-2025.11.27+c3799ba9e-py3-none-any.whl
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
strif==3.0.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
valkey==6.1.1
waitress==3.0.2
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.11.0
yarl==1.22.0
